requirements.txt
python-telegram-bot==20.6
pandas==2.1.1
numpy==1.26.2
matplotlib==3.8.0
scipy==1.11.2
